Add .gitignore for commonly ignored files
This commit introduces a .gitignore file to the repository. It's designed to ignore various types of files that are commonly not wanted in a git repo. These include system files for Linux, Windows, and macOS, as well as files related to several popular text editors (e.g., JetBrains, TextMate, Emacs). The ignored files are mostly temporary files, cache files, system files, and user-specific files that don't need to be shared among developers. This will help eliminate unnecessary files from being tracked and keep the repository clean and focused on project-specific files.